+++ projectBlueWater/src/main/java/com/lasthopesoftware/bluewater/client/browsing/items/access/CachedItemProvider.kt
@@ -11,11 +11,11 @@ import com.namehillsoftware.handoff.promises.Promise
 class CachedItemProvider(
 	private val inner: ProvideItems,
 	private val revisions: CheckRevisions,
-	private val itemFunctionCache: CachePromiseFunctions<Triple<LibraryId, ItemId?, Int>, List<Item>> = companionCache,
+	private val itemFunctionCache: CachePromiseFunctions<Triple<LibraryId, ItemId?, Long>, List<Item>> = companionCache,
 ) : ProvideItems {
 
 	companion object {
-		private val companionCache = LruPromiseCache<Triple<LibraryId, ItemId?, Int>, List<Item>>(20)
+		private val companionCache = LruPromiseCache<Triple<LibraryId, ItemId?, Long>, List<Item>>(20)
 	}
 
 	override fun promiseItems(libraryId: LibraryId, itemId: ItemId?): Promise<List<Item>> =
